+++ itella-cod.php
@@ -6,8 +6,8 @@
  * @package           Itella_COD_Plugin
  *
  * @wordpress-plugin
- * Plugin Name:       Smartpost Itella COD
- * Description:       Card on delivery payment method for Smartpost Itella shipping
+ * Plugin Name:       Smartposti COD
+ * Description:       Card on delivery payment method for Smartposti shipping
  * Version:           1.0.4
  * Author:            Itella Team
  * Author URI:        https://itella.lt/en/